fix: overflow error

+++ src/file/meta/smpte_offset.rs
@@ -90,9 +90,9 @@ impl SmpteOffset {
     /// # Returns
     /// The time offset in microseconds as a floating-point value
     pub const fn as_micros_with_override(&self, fps: SmpteFps) -> f64 {
-        ((((self.hour as u32 * 3600) + (self.minute as u32) * 60 + self.second as u32) * 1_000_000)
+        ((((self.hour as u64 * 3600) + (self.minute as u64) * 60 + self.second as u64) * 1_000_000)
             as f64)
-            + ((self.frame as u32) * 1_000_000) as f64 / fps.as_f64()
+            + ((self.frame as u64) * 1_000_000) as f64 / fps.as_f64()
             + ((self.subframe as u32) * 10_000) as f64 / fps.as_f64()
     }
     /// Convert this SMPTE offset to microseconds.
@@ -102,9 +102,9 @@ impl SmpteOffset {
     /// minutes, seconds, frames, and subframes to provide a precise
     /// microsecond value.
     pub const fn as_micros(&self) -> f64 {
-        ((((self.hour as u32 * 3600) + (self.minute as u32) * 60 + self.second as u32) * 1_000_000)
+        ((((self.hour as u64 * 3600) + (self.minute as u64) * 60 + self.second as u64) * 1_000_000)
             as f64)
-            + ((self.frame as u32) * 1_000_000) as f64 / self.fps.as_f64()
+            + ((self.frame as u64) * 1_000_000) as f64 / self.fps.as_f64()
             + ((self.subframe as u32) * 10_000) as f64 / self.fps.as_f64()
     }
